Create Visualizations with R | 1. Covid-19 Time series Plot
R is a free software environment for statistical computing and graphics, and is widely used by both academia and industry.  R software works on both Windows and Mac-OS. It was ranked no. 1 in a KDnuggets poll on top languages for analytics, data mining, and data science. RStudio is a user friendly environment for R that has become popular.  RCloud can be used for this  project.